move files to prompts directory
