+++ backend/core/prompts/schemas/knowledge_base.json
