+++ backend/core/prompts/templates/__init__.py
