+++ backend/core/prompts/templates/agents.yaml
