+++ backend/core/prompts/templates/base.yaml
