+++ backend/core/prompts/templates/browser.yaml
